add search controller
Done testing search by product name, modify attribute of product

+++ backend/src/database/migrations/20241017174540-create-cart_item.js
@@ -27,7 +27,7 @@ module.exports = {
         field: 'product_id',
       },
       price: {
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL(10, 3),
         allowNull: false,
         field: 'price',
       },
